% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en2grid.R
\name{en2grid}
\alias{en2grid}
\title{en2grid}
\usage{
en2grid(data, grid.size = 10)
}
\arguments{
\item{data}{Your data. This should be a dataframe with columns 'easting' and 'northing'. Additional columns are ignored.}

\item{grid.size}{The desired output grid size, in km. Currently, only 1 and 10 km grid cells are accepted.}
}
\value{
Grid cells exported as vector
}
\description{
Returns OSGB grid cell for OSGB easting/northing coordinates.
}
\examples{
test_data <- data.frame(easting = c(455328, 631490, 427622), northing = c(199154, 275050, 164165), grid10 = c('SU59', 'TM37', 'SU26'), grid1 = c('SU5599', 'TM3175', 'SU2764'))

grid2en(test_data$grid10, grid.size = 10)

en2grid(test_data, grid.size = 10)
}
