% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2en.R
\name{grid2en}
\alias{grid2en}
\title{grid2en}
\usage{
grid2en(grids, grid.size = 10)
}
\arguments{
\item{grids}{OSGB grid cells}

\item{grid.size}{Known size of the grid cells. Only valid options are 1 or 10 km grid cells.}
}
\value{
Gives a dataframe of easting and northing coordinates.
}
\description{
Returns OSGB easting/northing coordinates from OSGB grid cells.
}
\examples{
test_data <- data.frame(easting = c(455328, 631490, 427622), northing = c(199154, 275050, 164165), grid10 = c('SU59', 'TM37', 'SU26'), grid1 = c('SU5599', 'TM3175', 'SU2764'))

grid2en(test_data$grid10, grid.size = 10)

en2grid(test_data, grid.size = 10)
}
